Revert "Rewrite in Swift"
This reverts commit 763ec970038626cdcaf43c409b8b3c916627445b.

Unfortunately, it seems the `.executableTarget` is always built for the target platform,
but Xcode looks for a version built for macOS.
We could work around this by using a `.binaryTarget` instead, but that does not offer
a significant-enough benefit over just using a shell script.
This commit reverts to the previous approach, at least until a better solution is found.